[server] | refactor(): videos query pipe 수정
common에 위치했던 pipe를 videos 모듈에 포함했습니다. 동시에 조건문 로직을 한 데 정리했습니다.
