--- packages/server/src/common/pipes/get-videos.pipe/dto.ts
@@ -1,7 +1,0 @@
-export class GetVideosPipeDto {
-  public constructor(
-    public readonly page?: string,
-    public readonly sort?: string,
-    public readonly period?: string,
-  ) {}
-}
