+++ packages/server/src/video/dto/videos-request-query.dto.ts
@@ -1,0 +1,7 @@
+export class VideosRequestQueryDto {
+  public constructor(
+    public page: string,
+    public sort: string,
+    public period?: string,
+  ) {}
+}
